requirements.txt
fastapi[standard]
psycopg[binary]
python-jose
redis